refactor: more separation on video service functionalities.
