+++ server/src/modules/video_project/functions.js
@@ -1,21 +1,13 @@
-const {
-  defineFunction,
-  storeFile,
-  getFilePath,
-} = require("@modular-rest/server");
+const { defineFunction, getFilePath } = require("@modular-rest/server");
+
+const { extractGroupedSegments } = require("./media-processor.service");
+
+const { getVideoProjectModels } = require("./service");
 
 const {
-  exportVideoBySegments,
-  extractGroupedSegments,
-} = require("./mediaProcessor");
-
-const {
-  getVideoProjectModels,
-  findProjectById,
-  extractSegmentsWithFilePathFromProjectTimeline,
-} = require("./service");
-
-const timelineChain = require("../../chains/video-editor-grouped-segment-based");
+  generateTimelineByPrompt,
+  generateVideoRevision,
+} = require("./timeline.service");
 
 module.exports.functions = [
   // Generate the timeline based given prompt
@@ -23,59 +15,8 @@ module.exports.functions = [
   defineFunction({
     name: "generateTimeline",
     permissionTypes: ["user_access"],
-    callback: async ({ projectId, prompt, userId }) => {
-      if (!projectId) {
-        throw new Error("projectId is required.");
-      }
-
-      const { videoMediaModel, projectModel } = getVideoProjectModels();
-
-      const project = await projectModel
-        .findOne({
-          _id: projectId,
-          userId,
-        })
-        .exec();
-
-      if (!project) {
-        throw new Error("No project found with the given id");
-      }
-
-      const projectVideoMedia = await videoMediaModel
-        .find({ projectId })
-        .sort({ creation_time: 1 })
-        .exec();
-
-      let groupedSegments = [];
-
-      for (const videoMedia of projectVideoMedia) {
-        if (!videoMedia || !videoMedia.groupedSegments) {
-          continue;
-        }
-
-        const videoMediaJson = videoMedia.toObject();
-
-        groupedSegments = groupedSegments.concat(
-          videoMediaJson.groupedSegments.map((segment) => {
-            return {
-              ...segment,
-              processedVideoId: videoMediaJson._id.toString(),
-              fileId: videoMediaJson.fileId,
-            };
-          })
-        );
-      }
-
-      const timeline = await timelineChain.invoke({
-        editing_request: prompt,
-        grouped_segments: groupedSegments,
-      });
-
-      await projectModel.updateOne({ _id: projectId }, { $set: { timeline } });
-
-      return {
-        timeline,
-      };
+    callback: ({ projectId, prompt, userId }) => {
+      return generateTimelineByPrompt({ projectId, prompt, userId });
     },
   }),
 
@@ -84,80 +25,8 @@ module.exports.functions = [
   defineFunction({
     name: "generateVideoRevision",
     permissionTypes: ["user_access"],
-    callback: async ({ prompt, projectId, userId }) => {
-      if (!prompt || !projectId || !userId) {
-        throw new Error("prompt and ids is required. userId is required.");
-      }
-
-      const projectDoc = await findProjectById(projectId);
-
-      if (!projectDoc) {
-        throw new Error("No project found with the given id");
-      } else if (projectDoc.userId !== userId) {
-        throw new Error("User does not have access to the project");
-      }
-
-      // extract segments
-      const extractedSegments =
-        await extractSegmentsWithFilePathFromProjectTimeline(projectId);
-
-      //
-      // Save the video revision
-      //
-      const { videoRevisionModel } = getVideoProjectModels();
-
-      const newRevision = await videoRevisionModel.create({
-        userId,
-        projectId,
-        prompt,
-        isPending: true,
-        segments: extractedSegments,
-      });
-
-      //
-      // Export the video from the revision
-      //
-
-      exportVideoBySegments(extractedSegments)
-        .then(async (fileForSaving) => {
-          const exportedFileId = await storeFile({
-            file: fileForSaving,
-            ownerId: userId,
-            tag: newRevision._id.toString(),
-            removeFileAfterStore: true,
-          })
-            .then((savedFile) => savedFile._id.toString())
-            .catch((error) => "");
-
-          await videoRevisionModel
-            .updateOne(
-              {
-                _id: newRevision._id,
-              },
-              {
-                $set: { isPending: false, fileId: exportedFileId },
-              }
-            )
-            .exec()
-            .then((res) => {
-              console.log("Video revision exported successfully", res);
-            });
-        })
-        .catch(async (error) => {
-          console.error("Error exporting video revision", error);
-          await videoRevisionModel
-            .updateOne(
-              {
-                _id: newRevision._id,
-              },
-              {
-                $set: { isPending: false },
-              }
-            )
-            .exec();
-        });
-
-      return newRevision.toObject();
+    callback: ({ prompt, projectId, userId }) => {
+      return generateVideoRevision({ prompt, projectId, userId });
     },
   }),
